+++ jac3d_argobots/jac3d.c
@@ -34,7 +34,7 @@ static void configure_scheduler_mode(void) {
     g_use_cost_aware_scheduler = (strcmp(scheduler_mode, "new") == 0 || strcmp(scheduler_mode, "cost-aware") == 0);
 }
 
-static inline void register_task_estimate_if_needed(int pool_id, double estimate) {
+static inline void register_task_estimate_if_needed(int pool_id, long long estimate) {
     if (g_use_cost_aware_scheduler) {
         ws_push_task_estimate(pool_id, estimate);
     }
@@ -210,7 +210,9 @@ int main(int argc, char **argv) {
             thread_args[t].end_i = (t == num_chunks - 1) ? L - 1 : thread_args[t].start_i + rows_per_chunk;
             thread_args[t].eps_local = &eps_values[t];
             
-            register_task_estimate_if_needed(t % reduction_context.num_pools, (double)(rows_per_chunk * L * L));
+            register_task_estimate_if_needed(
+                t % reduction_context.num_pools,
+                (long long)rows_per_chunk * (long long)L * (long long)L);
             ABT_thread_create(
                 reduction_context.pools[t % reduction_context.num_pools],
                 update_A_thread,
@@ -228,7 +230,9 @@ int main(int argc, char **argv) {
         reduce_max_float(&reduction_context, eps_values, num_chunks, &eps);
         
         for (int t = 0; t < num_chunks; t++) {
-            register_task_estimate_if_needed(t % reduction_context.num_pools, (double)(rows_per_chunk * L * L));
+            register_task_estimate_if_needed(
+                t % reduction_context.num_pools,
+                (long long)rows_per_chunk * (long long)L * (long long)L);
             ABT_thread_create(
                 reduction_context.pools[t % reduction_context.num_pools],
                 update_B_thread,
