+++ workstealing_scheduler/abt_workstealing_scheduler_cost_aware.h
@@ -5,9 +5,9 @@
 void ABT_create_ws_scheds_cost_aware(int num, ABT_pool *pools, ABT_sched *scheds);
 
 /* Метаданные задач для выбора жертвы по оценочной стоимости очередей. */
-void ws_push_task_estimate(int rank, double est);
-double ws_pop_task_estimate(int rank);
-double ws_get_pool_estimated_load(int rank);
+void ws_push_task_estimate(int rank, long long est);
+long long ws_pop_task_estimate(int rank);
+long long ws_get_pool_estimated_load(int rank);
 void ws_reset_steal_count(void);
 long long ws_get_steal_count(void);
 long long ws_get_steal_ops_count(void);
